% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_cdfplot.R
\name{self_cdfplot}
\alias{self_cdfplot}
\title{self_cdfplot}
\usage{
self_cdfplot(adata, collist, main, xlab)
}
\arguments{
\item{adata}{adata is list}

\item{collist}{collist is atrray}

\item{main}{main is str}

\item{xlab}{ylab is str}
}
\value{
a figure
}
\description{
made by ZJRen 2021,7,23
}
\details{
cdfplot for myself

input:adata output:cdf
}
\examples{
a <- rnorm(n=100,mean = 0,sd=3);b <- rnorm(n=100,mean = 3,sd=3)
adata <- list(a=a,b=b)
self_cdfplot(adata = adata,collist = c("red","green"),main="",xlab="")

}
