% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_Heatmap.R
\name{self_Heatmap}
\alias{self_Heatmap}
\title{self_Heatmap}
\usage{
self_Heatmap(matrix_input, coltype)
}
\arguments{
\item{matrix_input}{Heatmap matrix}

\item{coltype}{coltype=c("YB","RB","BGR","GrayBlue")}
}
\value{
a figure
}
\description{
made by ZJRen 2021,7,23
}
\details{
Heatmap for myself

input:adata output:box
}
\examples{
matrix_test <- matrix(c(rnorm(100,mean = 1,sd = 2),rnorm(100,mean = 5,sd=3)),ncol = 5)
colnames(matrix_test) <- paste("rep",c(1:5),sep = "")
rownames(matrix_test) <- paste("gene",c(1:40),sep = "")
self_Heatmap(matrix_input=matrix_test,coltype="YB")

}
