% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_boxplot_withJitter.R
\name{selfboxplot_with_jitter}
\alias{selfboxplot_with_jitter}
\title{selfboxplot_with_jitter}
\usage{
selfboxplot_with_jitter(adata, ylab, main, collist, width)
}
\arguments{
\item{adata}{adata is list}

\item{ylab}{ylab is str}

\item{main}{main is str}

\item{collist}{collist is atrray}

\item{width}{width is point width}
}
\value{
a figure
}
\description{
made by ZJRen 2021,7,23
}
\details{
boxplot for myself

input:adata output:box
}
\examples{
adata=list(a=rnorm(n=10,mean=0,sd=1),b=rnorm(n=10,mean=2,sd=1),c=rnorm(n=10,mean=5,sd=1));
selfboxplot_with_jitter(adata=adata,ylab="value",main="main",collist=rainbow(3),width=0.2)


}
